some code refacturing remotes, wall switch and smart plug
updating the battery (alarm) reports, now making sure the alarm is
really off before sending the off command or really needs to go on.
kinda regret i added this for this reason (hard to determine for sure),
but don't want to remove a capability now.

fix power management notifications smart plug, were not going to trigger
in homey v1.3 and higher